backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
chromadb
pypdf
python-dotenv
openai
sentence-transformer